+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.10] - 2020-02-24
+### Added
+ - Support for authorized URLs (Unity Connect auto-login).
+ - Automatically log in the user when the URL for the shared game is clicked.
+ - Open the shared game automatically in the web browser when the upload is completed.
+
 ## [1.0.9] - 2019-12-13
 ### Changed
 - Upload metadata: use `applicationIdentifier` in `ProjectSettings.asset` for retrieving the Microgames's name.